+++ src/main/java/cloud/fogbow/ras/api/http/request/Cloud.java
@@ -1,6 +1,8 @@
-package cloud.fogbow.ras.api.http;
+package cloud.fogbow.ras.api.http.request;
 
 import cloud.fogbow.common.exceptions.FogbowException;
+import cloud.fogbow.ras.api.http.CommonKeys;
+import cloud.fogbow.ras.api.http.response.CloudList;
 import cloud.fogbow.ras.constants.ApiDocumentation;
 import cloud.fogbow.ras.constants.ConfigurationPropertyKeys;
 import cloud.fogbow.ras.constants.Messages;
@@ -28,7 +30,7 @@ public class Cloud {
 
     @ApiOperation(value = ApiDocumentation.Cloud.GET_OPERATION)
     @RequestMapping(method = RequestMethod.GET)
-    public ResponseEntity<List<String>> getCloudNamesNoMemberId(
+    public ResponseEntity<CloudList> getCloudNamesNoMemberId(
         @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
         @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
         throws FogbowException {
@@ -36,7 +38,7 @@ public class Cloud {
             LOGGER.info(Messages.Info.RECEIVING_GET_CLOUDS_REQUEST);
             String memberId = PropertiesHolder.getInstance().getProperty(ConfigurationPropertyKeys.LOCAL_MEMBER_ID_KEY);
             List<String> cloudNames = ApplicationFacade.getInstance().getCloudNames(memberId, federationTokenValue);
-            return new ResponseEntity<>(cloudNames, HttpStatus.OK);
+            return new ResponseEntity<>(new CloudList(cloudNames), HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
             throw e;
@@ -45,7 +47,7 @@ public class Cloud {
 
     @ApiOperation(value = ApiDocumentation.Cloud.GET_OPERATION_FOR_MEMBER)
     @RequestMapping(value = "/{memberId:.+}", method = RequestMethod.GET)
-    public ResponseEntity<List<String>> getCloudNames(
+    public ResponseEntity<CloudList> getCloudNames(
             @ApiParam(value = ApiDocumentation.CommonParameters.MEMBER_ID)
             @PathVariable String memberId,
             @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
@@ -54,7 +56,7 @@ public class Cloud {
         try {
             LOGGER.info(Messages.Info.RECEIVING_GET_CLOUDS_REQUEST);
             List<String> cloudNames = ApplicationFacade.getInstance().getCloudNames(memberId, federationTokenValue);
-            return new ResponseEntity<>(cloudNames, HttpStatus.OK);
+            return new ResponseEntity<>(new CloudList(cloudNames), HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
             throw e;
