+++ src/main/java/cloud/fogbow/ras/api/http/request/Network.java
@@ -1,12 +1,14 @@
-package cloud.fogbow.ras.api.http;
+package cloud.fogbow.ras.api.http.request;
 
 import cloud.fogbow.common.exceptions.*;
+import cloud.fogbow.ras.api.http.CommonKeys;
+import cloud.fogbow.ras.api.http.response.ResourceId;
 import cloud.fogbow.ras.constants.ApiDocumentation;
 import cloud.fogbow.ras.constants.Messages;
 import cloud.fogbow.ras.core.ApplicationFacade;
 import cloud.fogbow.ras.core.models.InstanceStatus;
 import cloud.fogbow.ras.core.models.ResourceType;
-import cloud.fogbow.ras.core.models.instances.PublicIpInstance;
+import cloud.fogbow.ras.core.models.instances.NetworkInstance;
 import cloud.fogbow.ras.core.models.securityrules.SecurityRule;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -20,100 +22,99 @@ import java.util.List;
 
 @CrossOrigin
 @RestController
-@RequestMapping(value = PublicIp.PUBLIC_IP_ENDPOINT)
-@Api(description = ApiDocumentation.PublicIp.API)
-public class PublicIp {
+@RequestMapping(value = Network.NETWORK_ENDPOINT)
+@Api(description = ApiDocumentation.Network.API)
+public class Network {
 
-    public static final String PUBLIC_IP_ENDPOINT = "publicIps";
-    public static final String ORDER_CONTROLLER_TYPE = "publicip";
+    public static final String NETWORK_ENDPOINT = "networks";
+    public static final String ORDER_CONTROLLER_TYPE = "network";
 
     public static final String SECURITY_RULES_ENDPOINT = "securityRules";
     public static final String SECURITY_RULE_NAME = "security rule";
 
-    private final Logger LOGGER = Logger.getLogger(PublicIp.class);
+    private final Logger LOGGER = Logger.getLogger(Network.class);
 
-    @ApiOperation(value = ApiDocumentation.PublicIp.CREATE_OPERATION)
+    @ApiOperation(value = ApiDocumentation.Network.CREATE_OPERATION)
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<String> createPublicIp(
-            @ApiParam(value = ApiDocumentation.PublicIp.CREATE_REQUEST_BODY)
-            @RequestBody cloud.fogbow.ras.api.parameters.PublicIp publicIp,
+    public ResponseEntity<ResourceId> createNetwork(
+            @ApiParam(value = ApiDocumentation.Network.CREATE_REQUEST_BODY)
+            @RequestBody cloud.fogbow.ras.api.parameters.Network network,
             @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
             @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
             throws FogbowException {
 
         try {
             LOGGER.info(String.format(Messages.Info.RECEIVING_CREATE_REQUEST, ORDER_CONTROLLER_TYPE));
-            String publicIpId = ApplicationFacade.getInstance().createPublicIp(publicIp.getOrder(), federationTokenValue);
-            return new ResponseEntity<String>(publicIpId, HttpStatus.CREATED);
+            String networkId = ApplicationFacade.getInstance().createNetwork(network.getOrder(), federationTokenValue);
+            return new ResponseEntity<>(new ResourceId(networkId), HttpStatus.CREATED);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
             throw e;
         }
     }
 
-    @ApiOperation(value = ApiDocumentation.PublicIp.GET_BY_ID_OPERATION)
-    @RequestMapping(value = "/{publicIpId}", method = RequestMethod.GET)
-    public ResponseEntity<PublicIpInstance> getPublicIp(
-            @ApiParam(value = ApiDocumentation.PublicIp.ID)
-            @PathVariable String publicIpId,
-            @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
-            @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
-            throws FogbowException {
-
-        try {
-            LOGGER.info(String.format(Messages.Info.RECEIVING_GET_REQUEST, ORDER_CONTROLLER_TYPE, publicIpId));
-            PublicIpInstance publicIpInstance =
-                ApplicationFacade.getInstance().getPublicIp(publicIpId, federationTokenValue);
-            return new ResponseEntity<>(publicIpInstance, HttpStatus.OK);
-        } catch (Exception e) {
-            LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
-            throw e;
-        }
-    }
-
-    @ApiOperation(value = ApiDocumentation.PublicIp.DELETE_OPERATION)
-    @RequestMapping(value = "/{publicIpId}", method = RequestMethod.DELETE)
-    public ResponseEntity<Boolean> deletePublicIp(
-            @ApiParam(value = ApiDocumentation.PublicIp.ID)
-            @PathVariable String publicIpId,
-            @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
-            @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
-            throws FogbowException {
-
-        try {
-            LOGGER.info(String.format(Messages.Info.RECEIVING_DELETE_REQUEST, ORDER_CONTROLLER_TYPE, publicIpId));
-            ApplicationFacade.getInstance().deletePublicIp(publicIpId, federationTokenValue);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (Exception e) {
-            LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
-            throw e;
-        }
-    }
-
-    @ApiOperation(value = ApiDocumentation.PublicIp.GET_OPERATION)
+    @ApiOperation(value = ApiDocumentation.Network.GET_OPERATION)
     @RequestMapping(value = "/status", method = RequestMethod.GET)
-    public ResponseEntity<List<InstanceStatus>> getAllPublicIpStatus(
+    public ResponseEntity<List<InstanceStatus>> getAllNetworksStatus(
             @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
             @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
             throws FogbowException {
 
         try {
             LOGGER.info(String.format(Messages.Info.RECEIVING_GET_ALL_REQUEST, ORDER_CONTROLLER_TYPE));
-            List<InstanceStatus> publicIpStatus =
-                ApplicationFacade.getInstance().getAllInstancesStatus(federationTokenValue, ResourceType.PUBLIC_IP);
-            return new ResponseEntity<>(publicIpStatus, HttpStatus.OK);
+            List<InstanceStatus> networkInstanceStatus =
+                ApplicationFacade.getInstance().getAllInstancesStatus(federationTokenValue, ResourceType.NETWORK);
+            return new ResponseEntity<>(networkInstanceStatus, HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
             throw e;
         }
     }
 
-    @ApiOperation(value = ApiDocumentation.PublicIp.CREATE_SECURITY_RULE_OPERATION)
-    @RequestMapping(value = "/{publicIpId}/" + SECURITY_RULES_ENDPOINT, method = RequestMethod.POST)
-    public ResponseEntity<String> createSecurityRule(
-            @ApiParam(value = ApiDocumentation.PublicIp.ID)
-            @PathVariable String publicIpId,
-            @ApiParam(value = ApiDocumentation.PublicIp.CREATE_SECURITY_RULE_REQUEST_BODY)
+    @ApiOperation(value = ApiDocumentation.Network.GET_BY_ID_OPERATION)
+    @RequestMapping(value = "/{networkId}", method = RequestMethod.GET)
+    public ResponseEntity<NetworkInstance> getNetwork(
+            @ApiParam(value = ApiDocumentation.Network.ID)
+            @PathVariable String networkId,
+            @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
+            @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
+            throws FogbowException {
+
+        try {
+            LOGGER.info(String.format(Messages.Info.RECEIVING_GET_REQUEST, ORDER_CONTROLLER_TYPE, networkId));
+            NetworkInstance networkInstance = ApplicationFacade.getInstance().getNetwork(networkId, federationTokenValue);
+            return new ResponseEntity<>(networkInstance, HttpStatus.OK);
+        } catch (Exception e) {
+            LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
+            throw e;
+        }
+    }
+
+    @ApiOperation(value = ApiDocumentation.Network.DELETE_OPERATION)
+    @RequestMapping(value = "/{networkId}", method = RequestMethod.DELETE)
+    public ResponseEntity<Boolean> deleteNetwork(
+            @ApiParam(value = ApiDocumentation.Network.ID)
+            @PathVariable String networkId,
+            @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
+            @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
+            throws FogbowException {
+
+        try {
+            LOGGER.info(String.format(Messages.Info.RECEIVING_DELETE_REQUEST, ORDER_CONTROLLER_TYPE, networkId));
+            ApplicationFacade.getInstance().deleteNetwork(networkId, federationTokenValue);
+            return new ResponseEntity<Boolean>(HttpStatus.OK);
+        } catch (Exception e) {
+            LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
+            throw e;
+        }
+    }
+
+    @ApiOperation(value = ApiDocumentation.Network.CREATE_SECURITY_RULE_OPERATION)
+    @RequestMapping(value = "/{networkId}/" + SECURITY_RULES_ENDPOINT, method = RequestMethod.POST)
+    public ResponseEntity<ResourceId> createSecurityRule(
+            @ApiParam(value = ApiDocumentation.Network.ID)
+            @PathVariable String networkId,
+            @ApiParam(value = ApiDocumentation.Network.CREATE_SECURITY_RULE_REQUEST_BODY)
             @RequestBody SecurityRule securityRule,
             @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
             @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
@@ -121,20 +122,20 @@ public class PublicIp {
 
         try {
             LOGGER.info(String.format(Messages.Info.RECEIVING_CREATE_REQUEST, SECURITY_RULE_NAME));
-            String ruleId = ApplicationFacade.getInstance().createSecurityRule(publicIpId, securityRule,
-                    federationTokenValue, ResourceType.PUBLIC_IP);
-            return new ResponseEntity<String>(ruleId, HttpStatus.CREATED);
+            String ruleId = ApplicationFacade.getInstance().createSecurityRule(networkId, securityRule,
+                    federationTokenValue, ResourceType.NETWORK);
+            return new ResponseEntity<>(new ResourceId(ruleId), HttpStatus.CREATED);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
             throw e;
         }
     }
 
-    @ApiOperation(value = ApiDocumentation.PublicIp.GET_SECURITY_RULE_OPERATION)
-    @RequestMapping(value = "/{publicIpId}/" + SECURITY_RULES_ENDPOINT, method = RequestMethod.GET)
+    @ApiOperation(value = ApiDocumentation.Network.GET_SECURITY_RULE_OPERATION)
+    @RequestMapping(value = "/{networkId}/" + SECURITY_RULES_ENDPOINT, method = RequestMethod.GET)
     public ResponseEntity<List<SecurityRule>> getAllSecurityRules(
-            @ApiParam(value = ApiDocumentation.PublicIp.ID)
-            @PathVariable String publicIpId,
+            @ApiParam(value = ApiDocumentation.Network.ID)
+            @PathVariable String networkId,
             @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
             @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
             throws FogbowException {
@@ -142,7 +143,7 @@ public class PublicIp {
         try {
             LOGGER.info(String.format(Messages.Info.RECEIVING_GET_ALL_REQUEST, SECURITY_RULE_NAME));
             List<SecurityRule> securityRules = ApplicationFacade.getInstance().
-                    getAllSecurityRules(publicIpId, federationTokenValue, ResourceType.PUBLIC_IP);
+                    getAllSecurityRules(networkId, federationTokenValue, ResourceType.NETWORK);
             return new ResponseEntity<>(securityRules, HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
@@ -150,12 +151,12 @@ public class PublicIp {
         }
     }
 
-    @ApiOperation(value = ApiDocumentation.PublicIp.DELETE_SECURITY_RULE_OPERATION)
-    @RequestMapping(value = "/{publicIpId}/" + SECURITY_RULES_ENDPOINT + "/{ruleId}", method = RequestMethod.DELETE)
+    @ApiOperation(value = ApiDocumentation.Network.DELETE_SECURITY_RULE_OPERATION)
+    @RequestMapping(value = "/{networkId}/" + SECURITY_RULES_ENDPOINT + "/{ruleId}", method = RequestMethod.DELETE)
     public ResponseEntity<Boolean> deleteSecurityRule(
-            @ApiParam(value = ApiDocumentation.PublicIp.ID)
-            @PathVariable String publicIpId,
-            @ApiParam(value = ApiDocumentation.PublicIp.SECURITY_RULE_ID)
+            @ApiParam(value = ApiDocumentation.Network.ID)
+            @PathVariable String networkId,
+            @ApiParam(value = ApiDocumentation.Network.SECURITY_RULE_ID)
             @PathVariable String ruleId,
             @ApiParam(value = ApiDocumentation.CommonParameters.FEDERATION_TOKEN)
             @RequestHeader(required = false, value = CommonKeys.FEDERATION_TOKEN_VALUE_HEADER_KEY) String federationTokenValue)
@@ -163,8 +164,8 @@ public class PublicIp {
 
         try {
             LOGGER.info(String.format(Messages.Info.RECEIVING_DELETE_REQUEST, SECURITY_RULE_NAME, ruleId));
-            ApplicationFacade.getInstance().deleteSecurityRule(publicIpId, ruleId, federationTokenValue,
-                    ResourceType.PUBLIC_IP);
+            ApplicationFacade.getInstance().deleteSecurityRule(networkId, ruleId, federationTokenValue,
+                    ResourceType.NETWORK);
             return new ResponseEntity<>(HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.info(String.format(Messages.Exception.GENERIC_EXCEPTION, e.getMessage()), e);
