+++ src/test/java/cloud/fogbow/ras/requests/api/local/http/ExceptionHandlerControllerTest.java
@@ -1,6 +1,6 @@
 package cloud.fogbow.ras.requests.api.local.http;
 
-import cloud.fogbow.ras.api.http.Compute;
+import cloud.fogbow.ras.api.http.request.Compute;
 import cloud.fogbow.ras.api.http.HttpExceptionToErrorConditionTranslator;
 import org.junit.Before;
 import org.mockito.Mockito;
