+++ src/test/java/cloud/fogbow/ras/requests/api/local/http/HttpExceptionToErrorConditionTranslatorTest.java
@@ -1,7 +1,7 @@
 package cloud.fogbow.ras.requests.api.local.http;
 
 import cloud.fogbow.ras.api.http.CommonKeys;
-import cloud.fogbow.ras.api.http.Compute;
+import cloud.fogbow.ras.api.http.request.Compute;
 import cloud.fogbow.ras.core.ApplicationFacade;
 import org.junit.Before;
 import org.junit.Test;
