+++ src/test/java/cloud/fogbow/ras/requests/api/local/http/ImageTest.java
@@ -3,7 +3,7 @@ package cloud.fogbow.ras.requests.api.local.http;
 import cloud.fogbow.common.exceptions.FogbowException;
 import cloud.fogbow.common.exceptions.InstanceNotFoundException;
 import cloud.fogbow.ras.api.http.CommonKeys;
-import cloud.fogbow.ras.api.http.Image;
+import cloud.fogbow.ras.api.http.request.Image;
 import cloud.fogbow.ras.core.ApplicationFacade;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
