+++ src/test/java/cloud/fogbow/ras/requests/api/local/http/PublicIpTest.java
@@ -2,7 +2,7 @@ package cloud.fogbow.ras.requests.api.local.http;
 
 import cloud.fogbow.common.exceptions.FogbowException;
 import cloud.fogbow.common.exceptions.InstanceNotFoundException;
-import cloud.fogbow.ras.api.http.Attachment;
+import cloud.fogbow.ras.api.http.request.Attachment;
 import cloud.fogbow.ras.api.http.CommonKeys;
 import cloud.fogbow.ras.core.ApplicationFacade;
 import cloud.fogbow.ras.core.models.InstanceStatus;
@@ -89,8 +89,9 @@ public class PublicIpTest {
 
         // verify
         int expectedStatus = HttpStatus.CREATED.value();
+        String expectedResponse = String.format("{\"id\":\"%s\"}", orderId);
         Assert.assertEquals(expectedStatus, result.getResponse().getStatus());
-        Assert.assertEquals(orderId, result.getResponse().getContentAsString());
+        Assert.assertEquals(expectedResponse, result.getResponse().getContentAsString());
 
         Mockito.verify(this.facade, Mockito.times(1))
                 .createAttachment(Mockito.any(AttachmentOrder.class), Mockito.anyString());
